pyproject.toml
[project]
name = "organizador-diplomacy"
version = "0.1.0"
description = "Organizador de partidas para la liga de Diplomacy.Ñol"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.7.0",
    "alembic>=1.14.0",
    "notion-client==3.0.0",
    "python-dotenv==1.2.2",
]

[dependency-groups]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "pyright==1.1.408",
    "ruff==0.15.7",
    "vulture==2.16",
]

[tool.vulture]
exclude = ["backend/test_*.py", "backend/**/test_*.py"]
ignore_decorators = ["@app.route", "@router.get", "@router.post", "@router.delete", "@router.put", "@router.patch", "@pytest.fixture", "@pytest_asyncio.fixture"]
ignore_names = ["main", "args", "kwargs", "conn", "client", "app", "pytestmark", "model_config", "settings", "catch_all", "__context"]
min_confidence = 80
paths = ["backend"]
make_whitelist = true

[tool.ruff]
# Rules for the fast linter
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "FBT", # flake8-boolean-trap
]
ignore = [
    "E501",   # Line too long (handled by formatter or ignored)
]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.pyright]
include = ["backend"]
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
pythonVersion = "3.13"
pythonPlatform = "Darwin"
venvPath = "."
venv = ".venv"

[project.scripts]
viewer      = "backend.main:app"
organizar   = "backend.organizador:organizar_partidas"
notion-sync = "backend.notion_sync:main"
